fix(ui): fix android back button not working
