+++ src/ui/pages/LockPage/LockPage.tsx
@@ -1,14 +1,17 @@
+import { App, AppState } from "@capacitor/app";
+import { Capacitor } from "@capacitor/core";
+import { Keyboard } from "@capacitor/keyboard";
 import i18n from "i18next";
 import { useEffect, useMemo, useState } from "react";
 import { useSelector } from "react-redux";
-import { Capacitor } from "@capacitor/core";
-import { Keyboard } from "@capacitor/keyboard";
-import { App, AppState } from "@capacitor/app";
 import { KeyStoreKeys, SecureStorage } from "../../../core/storage";
-import { useAppDispatch } from "../../../store/hooks";
+import { PublicRoutes, RoutePath } from "../../../routes/paths";
+import { useAppDispatch, useAppSelector } from "../../../store/hooks";
 import { getBiometricsCacheCache } from "../../../store/reducers/biometricsCache";
 import {
   geFirstAppLaunch,
+  getAuthentication,
+  getCurrentRoute,
   login,
   setFirstAppLaunch,
 } from "../../../store/reducers/stateCache";
@@ -19,20 +22,21 @@ import {
 } from "../../components/ErrorMessage";
 import { ForgotAuthInfo } from "../../components/ForgotAuthInfo";
 import { ForgotType } from "../../components/ForgotAuthInfo/ForgotAuthInfo.types";
-import { PageFooter } from "../../components/PageFooter";
-import { PasscodeModule } from "../../components/PasscodeModule";
-import { ResponsivePageLayout } from "../../components/layout/ResponsivePageLayout";
-import { useBiometricAuth } from "../../hooks/useBiometricsHook";
-import { useExitAppWithDoubleTap } from "../../hooks/exitAppWithDoubleTapHook";
-import "./LockPage.scss";
-import { BackEventPriorityType } from "../../globals/types";
 import {
   MaxLoginAttemptAlert,
   useLoginAttempt,
 } from "../../components/MaxLoginAttemptAlert";
+import { PageFooter } from "../../components/PageFooter";
+import { PasscodeModule } from "../../components/PasscodeModule";
+import { ResponsivePageLayout } from "../../components/layout/ResponsivePageLayout";
+import { BackEventPriorityType } from "../../globals/types";
+import { useExitAppWithDoubleTap } from "../../hooks/exitAppWithDoubleTapHook";
 import { usePrivacyScreen } from "../../hooks/privacyScreenHook";
+import { useBiometricAuth } from "../../hooks/useBiometricsHook";
+import "./LockPage.scss";
+import { showError } from "../../utils/error";
 
-const LockPage = () => {
+const LockPageContainer = () => {
   const pageId = "lock-page";
   const dispatch = useAppDispatch();
   const [passcode, setPasscode] = useState("");
@@ -156,6 +160,7 @@ const LockPage = () => {
       document.getElementById("passcode-button-1")?.focus();
     }
   }, []);
+  
   useEffect(() => {
     const handleAppStateChange = async (state: AppState) => {
       if (state.isActive) {
@@ -163,10 +168,10 @@ const LockPage = () => {
       }
     };
 
-    App.addListener("appStateChange", handleAppStateChange);
+    const listener = App.addListener("appStateChange", handleAppStateChange);
 
     return () => {
-      App.removeAllListeners();
+      listener.then((value) => value.remove()).catch((e) => showError("Unable to clear listener", e));
     };
   }, []);
 
@@ -234,5 +239,18 @@ const LockPage = () => {
     </ResponsivePageLayout>
   );
 };
+
+const LockPage = () => {
+  const currentRoute = useAppSelector(getCurrentRoute);
+  const authentication = useAppSelector(getAuthentication);
+
+  const isPublicPage = PublicRoutes.includes(currentRoute?.path as RoutePath);
+
+  if(isPublicPage || authentication.loggedIn) {
+    return null;
+  }
+
+  return <LockPageContainer />
+}
 
 export { LockPage };
